ANDROID: dm: Add android verity target
This device-mapper target is virtually a VERITY target. This
target is setup by reading the metadata contents piggybacked
to the actual data blocks in the block device. The signature
of the metadata contents are verified against the key included
in the system keyring. Upon success, the underlying verity
target is setup.

BUG: 27175947

Change-Id: Ie6b4d4d9591a09b751e5079d35f22c5f130e42e8
Signed-off-by: Badhri Jagan Sridharan <Badhri@google.com>
Git-commit: 86c9f22461e4fb0f3e6ce50b6c4ac5856e323230
Git-repo: https://git.linaro.org/kernel/linux-linaro-stable.git
Signed-off-by: Imran Khan <kimran@codeaurora.org>