QBT1000: copy qseecom handle to user when loading/unloading app
QBT1000 provides IOCTLs for loading and unloading a QSEE app.
In the input structure for these IOCTLs there is a pointer
to a qseecom handle which serves as an output parameter for
the IOCTLs. That is, the given handle (in client address space)
should be set to a valid handle value on load, and should be set
to 0 on unload.

The driver was missing a proper copy_to_user() call for this handle,
which sometimes resulted in unload not setting the handle to 0.

CRs-Fixed: 1059327
Signed-off-by: Abir Ghosh <abirg@codeaurora.org>
Change-Id: I31f205afb1f9bf0b6243e3f20f54022525c93b28