memshare: Add support to listen to SUBSYS_RAMDUMP_NOTIFICATION
Currently, on modem subsystem restart the area allocated from
HLOS to modem at runtime is not captured as part of the ramdump
collection. This patch creates separate elf files for all the
clients which are using memshare feature for memory allocation
which can be later used to load as dumps.

CRs-Fixed: 1058744
Change-Id: I66436b7c2d9e02658cfc4c195ca78852b09c01f7
Signed-off-by: Manoj Prabhu B <bmanoj@codeaurora.org>