power: smb135x-charger: Handle delay in src_detect state change
The SMB hardware does not generate an interrupt on the falling
edge of src_detect, there is only a change in its real-time
status which is detected by the driver while handling the usbin_uv
irq.  This can be a problem when the USB input voltage
drops slowly and software is unable to detect the falling-edge of
src_detect, thus failing to detect a USB removal.

Handle this by periodically checking for the src_detect falling
event after receiving the usbin_uv irq.

CRs-Fixed: 1077297
Change-Id: I6ec27cbf5efd07181df2f7a4c5af84f0e96724f2
Signed-off-by: Anirudh Ghayal <aghayal@codeaurora.org>