wlan: During unload do not flush eCsrForcedIbssLeave command
During unload if disconnect command is pending the command is not
flushed and del self sta is queued after this disconnect command.
Due to this the DEL BSS is sent before HAL STOP.

The logic to check if the command is disconnect does not handle
eCsrForcedIbssLeave command and thus DEL BSS is not sent before
HAL STOP if IBSS disconnect is in progress during unload.

To fix this add the case for eCsrForcedIbssLeave in logic to check
if command is disconnect.

Change-Id: I1dba859e4bc95761f5dc7c7160368ccdedbafa4d
CRs-Fixed: 2001956