wcnss: Fix getting iris id wrong for wcn3610 card
For wcn3610 and wcn3610v1, their IRIS card chip id are 0x9101
and 0x9110, so highest bit is 1. In wcnss_get_iris_name,
when iris_id shifts 16bits right, it can't match the right name
as the highest bit is filled up with 1.

Change type of iris_id from int to u32, and add mask for iris_id.

Change-Id: I3c17949fc327f9406326705191e4f5aedddfd3c4
CRs-Fixed: 2048990
Signed-off-by: Jingxiang Ge <jge@codeaurora.org>