msm: sde: Remove output fence object after user request completed
The fence object is allocated in the driver during the buffer queuing
time. When user requested the output buffer fence fd, driver should not
hold on to the fence object anymore because there is only a single
reference to the fence. When user destroy the fence by closing the fd,
the reference is cleared. Driver only needs to destroy the fence object
if the user does not request it.

CRs-Fixed: 2059181
Bug: 64442463
CVE-Fixed: CVE-2017-11031

Change-Id: Ic83d93fd3c7f404774007065df02b402adbf80af
Signed-off-by: Benjamin Chan <bkchan@codeaurora.org>