usb: gadget: Clear multi_req flag if UDC suspended
If a large request has to be queued onto an EP, then
multi_req flag for that EP is set before queueing the
request to the hardware. Now if the UDC is suspended during
cable disconnect before the request could be queued to the
HW, then multi_req flag is not cleared even by ep_disable.
Then on cable connect, re-enumeration fails with Large Request
error for that particular endpoint.

Fix this by clearing multi_req flag if the UDC suspend
happens before HW enqueue happens.

Change-Id: I77316192876b46ab82b775e01e2155309c2e91f8
Signed-off-by: Ajay Agarwal <ajaya@codeaurora.org>